+++ app/javascript/vue/tasks/data_attributes/field_synchronize/components/Field/FieldForm.vue
@@ -8,10 +8,16 @@
         :properties="attributes"
         v-model="selectedAttributes"
       />
-      <PredicateSelector
-        :predicates="predicates"
-        v-model="selectedPredicates"
-      />
+      <div class="horizontal-left-content gap-small align-end">
+        <PredicateSelector
+          :predicates="predicates"
+          v-model="selectedPredicates"
+        />
+        <ModalPredicates
+          :predicates="predicates"
+          @select="(item) => emit('user-predicate', item)"
+        />
+      </div>
     </template>
   </BlockLayout>
 </template>
@@ -20,6 +26,7 @@
 import BlockLayout from '@/components/layout/BlockLayout.vue'
 import PredicateSelector from './PredicateSelector.vue'
 import PropertySelector from './PropertySelector.vue'
+import ModalPredicates from './ModalPredicates.vue'
 
 defineProps({
   attributes: {
@@ -32,6 +39,8 @@ defineProps({
     default: () => []
   }
 })
+
+const emit = defineEmits(['user-predicate'])
 
 const selectedPredicates = defineModel('selectedPredicates', { type: Array })
 const selectedAttributes = defineModel('selectedAttributes', { type: Array })