+++ app/javascript/vue/tasks/data_attributes/field_synchronize/composables/useFieldSync.js
@@ -24,6 +24,7 @@ import {
 import Qs from 'qs'
 
 export function useFieldSync() {
+  const userPredicates = ref([])
   const attributes = ref([])
   const dataAttributes = ref([])
   const from = ref()
@@ -313,9 +314,9 @@ export function useFieldSync() {
     })
 
     request.then(({ body }) => {
+      body.index.push(...userPredicates.value)
       predicates.value = body.index.map((item) => {
-        const [id] = Object.keys(item)
-        const [name] = Object.values(item)
+        const [id, name] = Object.entries(item)[0]
 
         return {
           id: Number(id),
@@ -443,6 +444,7 @@ export function useFieldSync() {
   return {
     attributes,
     currentModel,
+    currentPage,
     extractOperation,
     filterUrl,
     from,
@@ -469,8 +471,9 @@ export function useFieldSync() {
     sortListByEmpty,
     sortListByMatched,
     tableList,
-    updatedCount,
+    to,
     totalUpdate,
-    to
+    updatedCount,
+    userPredicates
   }
 }